refactor(api-file-manager): move ExtractMetadataHandler and input interface to base package
Both S3 and server packages had identical handler and input types. The
handler now lives in the base package; provider packages re-export it.
Task implementations stay provider-specific (S3 vs filesystem).
