+++ docs/superpowers/specs/2026-06-23-handoff-file-manager-common-extraction.md
@@ -1,0 +1,33 @@
+# Session Handoff — 2026-06-23 — File Manager Common Code Extraction
+
+## What was done
+
+- Extracted all duplicated code from `api-file-manager-s3` and `api-file-manager-server` into the base `api-file-manager` package (12 commits, 9 tasks via subagent-driven development)
+- Created 3 DI abstractions (`GetUploadPayloadUseCase`, `CreateMultiPartUploadUseCase`, `CompleteMultiPartUploadUseCase`) so provider packages only implement storage-specific logic
+- Moved shared GraphQL upload schema (`FmUploadGraphQLSchema`) to base package — identical typeDefs, resolvers wired to DI abstractions
+- Moved shared utilities (checkPermissions, FileKey, FileNormalizer, FileUploadModifier, etc.), WriteFileMetadata feature, and asset delivery transformation utils to base package
+- Fixed path traversal vulnerability in CompleteMultiPartUpload — now validates fileKey segments and scopes containment to tenant root
+- All builds clean, 44 tests passing (39 in api-file-manager, 5 in api-file-manager-server)
+
+## Key decisions
+
+- `FileUploadModifierPlugin.type` unified from `"fm.s3.uploadModifier"` / `"fm.server.uploadModifier"` to `"fm.uploadModifier"` — breaking for external consumers, safe internally
+- Server upload use cases use `Request` abstraction from `@webiny/handler/abstractions/Request.js` for accessing FastifyRequest via DI (needed by `resolveServerUrl`)
+- Upload GraphQL resolvers in base package resolve `checkPermissions` and `createFileNormalizerFromContext` as plain function imports, not DI — they're utilities, not services
+- `PresignedPostPayloadDataResponse` type removed from both provider packages (dead code after refactoring) — the shared `UploadPayloadResponse` uses `Record<string, unknown>` for the `data` field since GraphQL types it as `JSON!`
+
+## Current state
+
+- Branch: `bruno/feat/api-file-manager-server`, 33 commits ahead of next (not pushed)
+- Tests: 44 passed (39 base + 5 server)
+- Build: all three packages passing
+- Unpushed commits: 33 (21 from prior api-file-manager-server implementation + 12 from this extraction)
+
+## What might come next
+
+- Push branch and create PR
+- Verify `app-file-manager` frontend works unchanged with both backends (same GraphQL API shape)
+- Pulumi infra for self-hosted deployment (ServiceManifest registration)
+- Integration tests with real Fastify server
+- Consider moving `ExtractMetadataHandler` to base (identical in both — only the task definition differs)
+- Consider moving `ExtractMetadataInput` interface to base to enable the above
