+++ packages/api-file-manager-server/src/features/ExtractMetadata/ExtractMetadataTask.ts
@@ -4,10 +4,7 @@ import { TaskDefinition } from "@webiny/api-core/features/task/TaskDefinition/in
 import { GlobalKeyValueStore } from "@webiny/api-core/features/keyValueStore/index.js";
 import { UpdateFileUseCase } from "@webiny/api-file-manager/features/file/UpdateFile/index.js";
 import { MetadataReader } from "@webiny/api-file-manager/features/upload/WriteFileMetadata/MetadataReader.js";
-
-export interface ExtractMetadataInput {
-    fileId: string;
-}
+import type { ExtractMetadataInput } from "@webiny/api-file-manager/features/extractMetadata/ExtractMetadataInput.js";
 
 class ExtractMetadataTaskImpl implements TaskDefinition.Interface<ExtractMetadataInput> {
     public readonly id = "fileManagerExtractMetadata";
