+++ packages/api-file-manager/src/features/extractMetadata/ExtractMetadataInput.ts
@@ -1,0 +1,3 @@
+export interface ExtractMetadataInput {
+    fileId: string;
+}
